optimize filtering by allowed 'From:' header
skip the step entirely if the allow/deny lists are empty, because in
that case there is no point in finding the header and parsing the
address.

+++ src/deliver.c
@@ -614,13 +614,15 @@ deliver_route_msg_list(connect_route *route, GList *msgout_list)
 		}
 
 		/* filter by allowed from header */
-		header *from_hdr = find_header(msgout->msg->hdr_list, HEAD_FROM);
-		if (from_hdr) {
-			if (!route_from_hdr_is_allowed(route, from_hdr->value)) {
-				DEBUG(6) debugf("from hdr `%s' is not allowed for this "
-						"route\n", from_hdr->value);
-				destroy_msg_out(msgout_cloned);
-				continue;
+		if (route->denied_from_hdrs || route->allowed_from_hdrs) {
+			header *from_hdr = find_header(msgout->msg->hdr_list, HEAD_FROM);
+			if (from_hdr) {
+				if (!route_from_hdr_is_allowed(route, from_hdr->value)) {
+					DEBUG(6) debugf("from hdr `%s' is not allowed for this route\n",
+					                from_hdr->value);
+					destroy_msg_out(msgout_cloned);
+					continue;
+				}
 			}
 		}
